+++ src/main/scala/edu/rit/cs/mmior/jsonoid/generator/Generator.scala
@@ -13,6 +13,8 @@ import edu.rit.cs.mmior.jsonoid.discovery.schemas.{
   NumberSchema,
   ObjectSchema,
   ProductSchema,
+  ReferenceObjectProperty,
+  ReferenceSchema,
   StringSchema
 }
 
@@ -31,17 +33,21 @@ object Generator {
     () => StringGenerator.generate(StringSchema())
   )
 
+  @SuppressWarnings(Array("org.wartremover.warts.Recursion"))
   def generateFromSchema(schema: JsonSchema[_]): JValue = {
     schema match {
       case s: ArraySchema   => ArrayGenerator.generate(s)
       case s: BooleanSchema => BooleanGenerator.generate(s)
       case s: EnumSchema    => EnumGenerator.generate(s)
       case s: IntegerSchema => IntegerGenerator.generate(s)
-      case s: NumberSchema  => NumberGenerator.generate(s)
       case s: NullSchema    => NullGenerator.generate(s)
+      case s: NumberSchema  => NumberGenerator.generate(s)
       case s: ObjectSchema  => ObjectGenerator.generate(s)
       case s: ProductSchema => ProductGenerator.generate(s)
-      case s: StringSchema  => StringGenerator.generate(s)
+      case s: ReferenceSchema =>
+        val refSchema = s.properties.get[ReferenceObjectProperty].schema
+        generateFromSchema(refSchema)
+      case s: StringSchema => StringGenerator.generate(s)
       case s: AnySchema =>
         AnyGenerators(Random.nextInt(AnyGenerators.size))()
     }